requirements.txt
pillow
numpy
soundfile